Fix the potential bugs in do_dom_coco_op
Adjust the signature of do_dom_coco_op to be aligned
with how other hypercall signatures are implemented.
Handle all arguments via XEN_GUEST_HANDLE_PARAM and
copy the arguments struct using copy_from_guest.

Also, add the update the declaration in hypercall-defs
and fix some casting issues

+++ tools/libs/light/libxl_dom.c
@@ -1146,7 +1146,7 @@ int libxl__build_hvm(libxl__gc *gc, uint32_t domid,
 
     if (info->coco.val) {
         LOG(DEBUG, "xc_dom_coco_op called, domid is %d", domid);
-        rc = xc_dom_coco_op(ctx->xch, 1, domid, (uint64_t)dom->kernel_blob, dom->kernel_size);
+        rc = xc_dom_coco_op(ctx->xch, 1, (uint16_t)domid, (uint64_t)dom->kernel_blob, dom->kernel_size);
         if (rc) {
             LOG(ERROR, "Failed to encrypt kernel memory for domain\n");
             goto out;
